requirements.txt
fastapi[standard]
fastapi-mail
sqlmodel==0.0.23
pyodbc
PyJWT
cryptocode
python-dotenv
insightface
onnxruntime
Pillow
numpy